refac: initialData 분리
