+++ src/basic/App.tsx
@@ -1,16 +1,12 @@
 import { useState, useCallback, useEffect } from "react";
-import { CartItem, Coupon, Product } from "./types";
+import { CartItem, Coupon, Product, ProductWithUI } from "./types";
 import {
   formatPriceAtCart,
   formatPriceAtAdmin,
   formatPercentage,
 } from "./utils/formatters";
-
-// 기본 Product에 UI 관련 속성을 추가한 확장 타입
-interface ProductWithUI extends Product {
-  description?: string;
-  isRecommended?: boolean;
-}
+import initialProducts from "./data/initialProducts";
+import initialCoupons from "./data/inintialCoupons";
 
 // 알림 메시지 타입
 interface Notification {
@@ -18,57 +14,6 @@ interface Notification {
   message: string;
   type: "error" | "success" | "warning";
 }
-
-// 📦 상품 초기 데이터 - 실제로는 API에서 가져올 데이터
-const initialProducts: ProductWithUI[] = [
-  {
-    id: "p1",
-    name: "상품1",
-    price: 10000,
-    stock: 20,
-    discounts: [
-      { quantity: 10, rate: 0.1 },
-      { quantity: 20, rate: 0.2 },
-    ],
-    description: "최고급 품질의 프리미엄 상품입니다.",
-  },
-  {
-    id: "p2",
-    name: "상품2",
-    price: 20000,
-    stock: 20,
-    discounts: [{ quantity: 10, rate: 0.15 }],
-    description: "다양한 기능을 갖춘 실용적인 상품입니다.",
-    isRecommended: true,
-  },
-  {
-    id: "p3",
-    name: "상품3",
-    price: 30000,
-    stock: 20,
-    discounts: [
-      { quantity: 10, rate: 0.2 },
-      { quantity: 30, rate: 0.25 },
-    ],
-    description: "대용량과 고성능을 자랑하는 상품입니다.",
-  },
-];
-
-// 🎫 쿠폰 초기 데이터
-const initialCoupons: Coupon[] = [
-  {
-    name: "5000원 할인",
-    code: "AMOUNT5000",
-    discountType: "amount",
-    discountValue: 5000,
-  },
-  {
-    name: "10% 할인",
-    code: "PERCENT10",
-    discountType: "percentage",
-    discountValue: 10,
-  },
-];
 
 const App = () => {
   // ========== 📋 상태 관리 섹션 ==========
