+++ src/basic/data/inintialCoupons.ts
@@ -1,0 +1,18 @@
+import { Coupon } from "../types";
+
+const initialCoupons: Coupon[] = [
+  {
+    name: "5000원 할인",
+    code: "AMOUNT5000",
+    discountType: "amount",
+    discountValue: 5000,
+  },
+  {
+    name: "10% 할인",
+    code: "PERCENT10",
+    discountType: "percentage",
+    discountValue: 10,
+  },
+];
+
+export default initialCoupons;
