+++ src/basic/types.ts
@@ -6,6 +6,11 @@ export interface Product {
   discounts: Discount[];
 }
 
+export interface ProductWithUI extends Product {
+  description?: string;
+  isRecommended?: boolean;
+}
+
 export interface Discount {
   quantity: number;
   rate: number;
@@ -19,6 +24,6 @@ export interface CartItem {
 export interface Coupon {
   name: string;
   code: string;
-  discountType: 'amount' | 'percentage';
+  discountType: "amount" | "percentage";
   discountValue: number;
 }
